pyproject.toml
[project]
name = "video-analysis-bot"
version = "0.1.0"
description = "RLT тестовое задание"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.23.0",
    "loguru>=0.7.3",
    "openai>=2.12.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "tortoise-orm[asyncpg]>=0.25.2",
]

[tool.pyright]
executionEnvironments = [{ root = "video_analysis_bot" }]
venvPath = "."
venv = ".venv"
